# Conflicts:
#	engine/GPEngine.vcxproj
#	engine/GPEngine.vcxproj.filters
#	projects/GPGame/include/Game.hpp
#	projects/GPGame/include/myScript.hpp
#	projects/GPGame/src/game.cpp

+++ engine/include/Engine/ECS/Component/Camera.hpp
@@ -84,9 +84,9 @@ protected:
     void updateProjection();
 
 public:
-    static float computeAspect(float width, float height) noexcept
+    static float computeAspect(int width, int height) noexcept
     {
-        return width / height;
+        return width / static_cast<float>(height);
     }
 
 public:
